% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tables_db_fxns.R
\name{connect_to_tables}
\alias{connect_to_tables}
\title{Connect to duckdb}
\usage{
connect_to_tables(db_path)
}
\arguments{
\item{db_path}{path to duckdb}
}
\value{
db connection
}
\description{
Connect to duckdb
}
