% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tables_db_fxns.R
\name{create_conditions}
\alias{create_conditions}
\title{Create conditions}
\usage{
create_conditions(...)
}
\arguments{
\item{...}{}
}
\value{
list of conditions
}
\description{
Creates a list of conditions to pass to query_tables functions
}
