% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_annualized.R
\name{query_annualized}
\alias{query_annualized}
\title{Query annualized estimates of measurements}
\usage{
query_annualized(
  con,
  conditions = create_conditions(...),
  variables = c("DIA")
)
}
\arguments{
\item{con}{database connection}

\item{conditions}{use \code{create_conditions} to create a list of conditions involving columns from the tree, plot, cond, qa_flags, or tree_info_composite_id tables}

\item{variables}{character vector of variables to return}
}
\value{
data frame of annual measurements and additional varaibles estimated for trees satisfying conditions
}
\description{
Pulls timeseries of \emph{annual} measurements interpolated for trees.
}
