% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_saplings.R
\name{query_saplings}
\alias{query_saplings}
\title{Pull records from the sapling transition tables}
\usage{
query_saplings(
  con,
  conditions = create_conditions(...),
  variables = c("STATECD")
)
}
\arguments{
\item{con}{database connection}

\item{conditions}{use \code{create_conditions} to create a list of logical conditions involving columns from the PLOT table}

\item{variables}{character vector or variables to return}
}
\value{
data frame of sapling transitions + any additional variables for plots satisfying conditions
}
\description{
Pull records from the sapling transition tables
}
