% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tables_db_fxns.R
\name{query_tree_surveys}
\alias{query_tree_surveys}
\title{Query tree surveys}
\usage{
query_tree_surveys(
  con,
  conditions = create_conditions(...),
  variables = c("DIA")
)
}
\arguments{
\item{con}{database connection}

\item{conditions}{use \code{create_conditions} to create a set of conditions}

\item{variables}{character vector of variables to pull}
}
\value{
data frame of variables for all tree records satisfying conditions
}
\description{
Pull timeseries of records for individual trees.
You can filter based on any logical conditions involving columns from the TREE, PLOT, COND, qa_flags, or tree_info_composite_id tables.
}
