% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_areas}
\alias{state_areas}
\title{State codes used in FIA}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 50 rows and 3 columns.
}
\usage{
state_areas
}
\description{
A data frame tha matches the \code{STATECD} column used in FIA data to the two
letter state abbreviations and the total land area of the state in acres.
}
\keyword{datasets}
